[#156742710] Update the AWS iOS SDK to allow us to provide a pre-signed URL for the AWSS3TransferUtility to use when performing common tasks |
WIP Adding the overloads to allow the preSignedURL to be provided instead of generating. Given how the multipart uploading works, we need to re-configure the transfer utility to allow for a provider of preSignedURLs to be associated so that each multipart preSignedURL can be requested by the transfer utility itself.